fix getUserDataContext 정리
