+++ src/context/myWekiDataContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { userType } from '@api/user';
+
+export const MyWekiDataContext = createContext<userType | null>(null);
